Separated dialogs which deal with addresses into address_dialogs module
